requirements.txt
jacobs-jinja-too
yamlloader
pyyaml